ZipMaster-PQre 熊云熠修改 1.0/ZipMaster-PQre/requirements.txt
# -*- coding: utf-8 -*-
# 压缩文件处理
py7zr>=0.20.0
rarfile>=4.0
patool>=1.12.0

# 文件监控
watchdog>=3.0.0

# GUI 增强 (可选)
# PyQt6>=6.5.0
# 或者使用内置的 tkinter

# 开发工具
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
